+++ api/admin.py
@@ -3,6 +3,7 @@ from django.utils.translation import gettext_lazy as _
 from django.utils.html import format_html_join
 from django.utils.safestring import mark_safe
 from api.event_sources import SOURCES
+from api.admin_classes import RegionRestrictedAdmin
 import api.models as models
 
 
@@ -102,6 +103,10 @@ class EventSourceFilter(admin.SimpleListFilter):
             return queryset.filter(auto_generated=True).filter(auto_generated_source__isnull=True)
 
 
+class DisasterTypeAdmin(admin.ModelAdmin):
+    search_fields = ('name',)
+
+
 class KeyFigureInline(admin.TabularInline):
     model = models.KeyFigure
 
@@ -118,11 +123,14 @@ class SituationReportInline(admin.TabularInline):
     model = models.SituationReport
 
 
-class EventAdmin(admin.ModelAdmin):
+class EventAdmin(RegionRestrictedAdmin):
+    country_in = 'countries__pk__in'
+    region_in = 'regions__pk__in'
+
     inlines = [KeyFigureInline, SnippetInline, EventContactInline, SituationReportInline]
     list_display = ('name', 'alert_level', 'glide', 'auto_generated', 'auto_generated_source',)
     list_filter = [IsFeaturedFilter, EventSourceFilter,]
-    search_fields = ['name', 'countries', 'dtype',]
+    search_fields = ('name', 'countries__name', 'dtype__name',)
     readonly_fields = ('appeals', 'field_reports', 'auto_generated_source',)
     autocomplete_fields = ('countries', 'districts',)
     def appeals(self, instance):
@@ -146,6 +154,12 @@ class EventAdmin(admin.ModelAdmin):
     field_reports.short_description = 'Field Reports'
 
 
+class GdacsAdmin(RegionRestrictedAdmin):
+    country_in = 'countries__pk__in'
+    region_in = None
+    search_fields = ('title',)
+
+
 class ActionsTakenInline(admin.TabularInline):
     model = models.ActionsTaken
 
@@ -158,14 +172,17 @@ class FieldReportContactInline(admin.TabularInline):
     model = models.FieldReportContact
 
 
-class FieldReportAdmin(admin.ModelAdmin):
+class FieldReportAdmin(RegionRestrictedAdmin):
+    country_in = 'countries__pk__in'
+    region_in = 'regions__pk__in'
+
     inlines = [ActionsTakenInline, SourceInline, FieldReportContactInline]
-    list_display = ('summary', 'user', 'event', 'visibility',)
-    list_editable = ('event',)
+    list_display = ('summary', 'event', 'visibility',)
     list_select_related = ('event',)
-    search_fields = ['countries', 'regions', 'summary',]
-    autocomplete_fields = ('countries', 'districts',)
-    list_filter = [HasRelatedEventFilter, MembershipFilter,]
+    search_fields = ('countries', 'regions', 'summary',)
+    autocomplete_fields = ('event', 'countries', 'districts',)
+    readonly_fields = ('report_date', 'created_at', 'updated_at',)
+    list_filter = [MembershipFilter,]
     actions = ['create_events',]
 
     def create_events(self, request, queryset):
@@ -193,15 +210,17 @@ class AppealDocumentInline(admin.TabularInline):
     model = models.AppealDocument
 
 
-class AppealAdmin(admin.ModelAdmin):
+class AppealAdmin(RegionRestrictedAdmin):
+    country_in = 'country__pk__in'
+    region_in = 'region__pk__in'
     inlines = [AppealDocumentInline]
     list_display = ('code', 'name', 'atype', 'needs_confirmation', 'event', 'start_date',)
-    list_editable = ('event',)
     list_select_related = ('event',)
-    search_fields = ['code', 'name',]
+    search_fields = ('code', 'name',)
     readonly_fields = ('region',)
     list_filter = [HasRelatedEventFilter, AppealTypeFilter,]
     actions = ['create_events', 'confirm_events',]
+    autocomplete_fields = ('event', 'country',)
 
     def create_events(self, request, queryset):
         for appeal in queryset:
@@ -241,6 +260,12 @@ class AppealAdmin(admin.ModelAdmin):
         super().save_model(request, obj, form, change)
 
 
+class AppealDocumentAdmin(RegionRestrictedAdmin):
+    country_in = 'appeal__country__in'
+    region_in = 'appeal__region__in'
+    search_fields = ('name', 'appeal__code', 'appeal__name')
+
+
 class CountryKeyFigureInline(admin.TabularInline):
     model = models.CountryKeyFigure
 
@@ -273,27 +298,49 @@ class RegionContactInline(admin.TabularInline):
     model = models.RegionContact
 
 
-class DistrictAdmin(admin.ModelAdmin):
+class DistrictAdmin(RegionRestrictedAdmin):
+    country_in = 'country__pk__in'
+    region_in = 'country__region__in'
     search_fields = ('name', 'country_name',)
 
 
-class CountryAdmin(admin.ModelAdmin):
+class CountryAdmin(RegionRestrictedAdmin):
+    country_in = 'pk__in'
+    region_in = 'region__pk__in'
     search_fields = ('name',)
     inlines = [CountryKeyFigureInline, CountrySnippetInline, CountryLinkInline, CountryContactInline,]
 
 
-class RegionAdmin(admin.ModelAdmin):
+class RegionAdmin(RegionRestrictedAdmin):
+    country_in = None
+    region_in = 'pk__in'
     inlines = [RegionKeyFigureInline, RegionSnippetInline, RegionLinkInline, RegionContactInline,]
+    search_fields = ('name',)
 
 
-admin.site.register(models.DisasterType)
+class UserProfileAdmin(admin.ModelAdmin):
+    search_fields = ('user__username', 'user__email',)
+
+
+class SituationReportAdmin(RegionRestrictedAdmin):
+    country_in = 'event__countries__in'
+    region_in = 'event__regions__in'
+    search_fields = ('name', 'event__name',)
+
+
+class SituationReportTypeAdmin(admin.ModelAdmin):
+    search_fields = ('type',)
+
+
+admin.site.register(models.DisasterType, DisasterTypeAdmin)
 admin.site.register(models.Event, EventAdmin)
-admin.site.register(models.GDACSEvent)
+admin.site.register(models.GDACSEvent, GdacsAdmin)
 admin.site.register(models.Country, CountryAdmin)
 admin.site.register(models.Region, RegionAdmin)
 admin.site.register(models.District, DistrictAdmin)
 admin.site.register(models.Appeal, AppealAdmin)
-admin.site.register(models.AppealDocument)
+admin.site.register(models.AppealDocument, AppealDocumentAdmin)
 admin.site.register(models.FieldReport, FieldReportAdmin)
-admin.site.register(models.Profile)
-admin.site.register(models.SituationReport)
+admin.site.register(models.Profile, UserProfileAdmin)
+admin.site.register(models.SituationReport, SituationReportAdmin)
+admin.site.register(models.SituationReportType, SituationReportTypeAdmin)