+++ api/drf_views.py
@@ -1,9 +1,14 @@
+from rest_framework.status import HTTP_201_CREATED, HTTP_200_OK
+from rest_framework.generics import GenericAPIView, CreateAPIView, UpdateAPIView
+from rest_framework.response import Response
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
 from rest_framework import viewsets
 from django_filters import rest_framework as filters
 from django.contrib.auth.models import User, Group
+from .event_sources import SOURCES
+from .exceptions import BadRequest
+from .view_filters import ListFilter
 from .models import (
     DisasterType,
 
@@ -17,13 +22,22 @@ from .models import (
 
     District,
 
+    Snippet,
     Event,
     Snippet,
     SituationReport,
+    SituationReportType,
     Appeal,
     AppealDocument,
     Profile,
     FieldReport,
+    FieldReportContact,
+    ActionsTaken,
+    Source,
+    SourceType,
+
+    VisibilityChoices,
+    RequestChoices,
 )
 
 from .serializers import (
@@ -43,16 +57,20 @@ from .serializers import (
     MiniDistrictSerializer,
     SnippetSerializer,
 
+    SnippetSerializer,
     ListEventSerializer,
     DetailEventSerializer,
     SituationReportSerializer,
+    SituationReportTypeSerializer,
     AppealSerializer,
     AppealDocumentSerializer,
     UserSerializer,
     ProfileSerializer,
     ListFieldReportSerializer,
     DetailFieldReportSerializer,
+    CreateFieldReportSerializer,
 )
+from .logger import logger
 
 class DisasterTypeViewset(viewsets.ReadOnlyModelViewSet):
     queryset = DisasterType.objects.all()
@@ -65,8 +83,15 @@ class RegionViewset(viewsets.ReadOnlyModelViewSet):
             return RegionSerializer
         return RegionRelationSerializer
 
+class CountryFilter(filters.FilterSet):
+    region = filters.NumberFilter(name='region', lookup_expr='exact')
+    class Meta:
+        model = Country
+        fields = ('region',)
+
 class CountryViewset(viewsets.ReadOnlyModelViewSet):
     queryset = Country.objects.all()
+    filter_class = CountryFilter
     def get_serializer_class(self):
         if self.action == 'list':
             return CountrySerializer
@@ -148,6 +173,9 @@ class DistrictViewset(viewsets.ReadOnlyModelViewSet):
 class EventFilter(filters.FilterSet):
     dtype = filters.NumberFilter(name='dtype', lookup_expr='exact')
     is_featured = filters.BooleanFilter(name='is_featured')
+    countries__in = ListFilter(name='countries__id')
+    regions__in = ListFilter(name='regions__id')
+    id = filters.NumberFilter(name='id', lookup_expr='exact')
     class Meta:
         model = Event
         fields = {
@@ -162,12 +190,19 @@ class EventViewset(viewsets.ReadOnlyModelViewSet):
             return ListEventSerializer
         else:
             return DetailEventSerializer
-    ordering_fields = ('disaster_start_date', 'created_at', 'name', 'summary', 'num_affected',)
+    ordering_fields = ('disaster_start_date', 'created_at', 'name', 'summary', 'num_affected', 'glide', 'alert_level',)
     filter_class = EventFilter
 
+class EventSnippetFilter(filters.FilterSet):
+    event = filters.NumberFilter(name='event', lookup_expr='exact')
+    class Meta:
+        model = Snippet
+        fields = ('event',)
+
 class EventSnippetViewset(viewsets.ReadOnlyModelViewSet):
-    serializer_class = SnippetSerializer
     authentication_classes = (TokenAuthentication,)
+    serializer_class = SnippetSerializer
+    filter_class = EventSnippetFilter
     def get_queryset(self):
         user_groups = self.request.user.groups.all()
         ifrc_group = Group.objects.get(pk=2)
@@ -178,9 +213,35 @@ class EventSnippetViewset(viewsets.ReadOnlyModelViewSet):
                 return Snippet.objects.exclude(visibility=2)
         return Snippet.objects.filter(visibility=3)
 
+class CountrySnippetViewset(viewsets.ReadOnlyModelViewSet):
+    authentication_classes = (TokenAuthentication,)
+    serializer_class = CountrySnippetSerializer
+    filter_class = CountrySnippetFilter
+    def get_queryset(self):
+        if self.request.user.is_authenticated:
+            return CountrySnippet.objects.all()
+        return CountrySnippet.objects.filter(visibility=3)
+
+class SituationReportTypeViewset(viewsets.ReadOnlyModelViewSet):
+    queryset = SituationReportType.objects.all()
+    serializer_class = SituationReportTypeSerializer
+    ordering_fields = ('type',)
+
+class SituationReportFilter(filters.FilterSet):
+    event = filters.NumberFilter(name='event', lookup_expr='exact')
+    type = filters.NumberFilter(name='type', lookup_expr='exact')
+    class Meta:
+        model = SituationReport
+        fields = {
+            'name': ('exact',),
+            'created_at': ('exact', 'gt', 'gte', 'lt', 'lte'),
+        }
+
 class SituationReportViewset(viewsets.ReadOnlyModelViewSet):
     queryset = SituationReport.objects.all()
     serializer_class = SituationReportSerializer
+    ordering_fields = ('created_at', 'name',)
+    filter_class = SituationReportFilter
 
 class AppealFilter(filters.FilterSet):
     atype = filters.NumberFilter(name='atype', lookup_expr='exact')
@@ -188,6 +249,8 @@ class AppealFilter(filters.FilterSet):
     country = filters.NumberFilter(name='country', lookup_expr='exact')
     region = filters.NumberFilter(name='region', lookup_expr='exact')
     code = filters.CharFilter(name='code', lookup_expr='exact')
+    status = filters.NumberFilter(name='status', lookup_expr='exact')
+    id = filters.NumberFilter(name='id', lookup_expr='exact')
     class Meta:
         model = Appeal
         fields = {
@@ -198,7 +261,7 @@ class AppealFilter(filters.FilterSet):
 class AppealViewset(viewsets.ReadOnlyModelViewSet):
     queryset = Appeal.objects.all()
     serializer_class = AppealSerializer
-    ordering_fields = ('start_date', 'end_date', 'name', 'aid', 'dtype', 'num_beneficiaries', 'amount_requested', 'amount_funded',)
+    ordering_fields = ('start_date', 'end_date', 'name', 'aid', 'dtype', 'num_beneficiaries', 'amount_requested', 'amount_funded', 'status', 'atype', 'event',)
     filter_class = AppealFilter
 
     def remove_unconfirmed_event(self, obj):
@@ -226,9 +289,21 @@ class AppealViewset(viewsets.ReadOnlyModelViewSet):
         serializer = self.get_serializer(instance)
         return Response(self.remove_unconfirmed_event(serializer.data))
 
+class AppealDocumentFilter(filters.FilterSet):
+    appeal = filters.NumberFilter(name='appeal', lookup_expr='exact')
+    appeal__in = ListFilter(name='appeal__id')
+    class Meta:
+        model = AppealDocument
+        fields = {
+            'name': ('exact',),
+            'created_at': ('exact', 'gt', 'gte', 'lt', 'lte'),
+        }
+
 class AppealDocumentViewset(viewsets.ReadOnlyModelViewSet):
     queryset = AppealDocument.objects.all()
     serializer_class = AppealDocumentSerializer
+    ordering_fields = ('created_at', 'name',)
+    filter_class = AppealDocumentFilter
 
 class ProfileViewset(viewsets.ModelViewSet):
     serializer_class = ProfileSerializer
@@ -247,6 +322,9 @@ class UserViewset(viewsets.ModelViewSet):
 class FieldReportFilter(filters.FilterSet):
     dtype = filters.NumberFilter(name='dtype', lookup_expr='exact')
     user = filters.NumberFilter(name='user', lookup_expr='exact')
+    countries__in = ListFilter(name='countries__id')
+    regions__in = ListFilter(name='regions__id')
+    id = filters.NumberFilter(name='id', lookup_expr='exact')
     class Meta:
         model = FieldReport
         fields = {
@@ -254,7 +332,7 @@ class FieldReportFilter(filters.FilterSet):
             'updated_at': ('exact', 'gt', 'gte', 'lt', 'lte'),
         }
 
-class FieldReportViewset(viewsets.ModelViewSet):
+class FieldReportViewset(viewsets.ReadOnlyModelViewSet):
     authentication_classes = (TokenAuthentication,)
     def get_queryset(self):
         if self.request.user.is_authenticated:
@@ -268,5 +346,252 @@ class FieldReportViewset(viewsets.ModelViewSet):
         else:
             return DetailFieldReportSerializer
 
-    ordering_fields = ('summary', 'created_at', 'updated_at')
+    ordering_fields = ('summary', 'event', 'dtype', 'created_at', 'updated_at')
     filter_class = FieldReportFilter
+
+class GenericFieldReportView(GenericAPIView):
+    authentication_classes = (TokenAuthentication,)
+    permission_classes = (IsAuthenticated,)
+    queryset = FieldReport.objects.all()
+
+    def serialize(self, data, instance=None):
+        # Replace integer values for Int Enum types.
+        # Otherwise, validation will fail.
+        # This applies to visibility and request choices.
+        if data['visibility'] == 2 or data['visibility'] == '2':
+            data['visibility'] = VisibilityChoices.IFRC
+        elif data['visibility'] == 3 or data['visibility'] == '3':
+            data['visibility'] = VisibilityChoices.PUBLIC
+        else:
+            data['visibility'] = VisibilityChoices.MEMBERSHIP
+
+        request_choices = [
+            'bulletin',
+            'dref',
+            'appeal',
+            'rdrt',
+            'fact',
+            'ifrc_staff',
+            'eru_base_camp',
+            'eru_basic_health_care',
+            'eru_it_telecom',
+            'eru_logistics',
+            'eru_deployment_hospital',
+            'eru_referral_hospital',
+            'eru_relief',
+            'eru_water_sanitation_15',
+            'eru_water_sanitation_40',
+            'eru_water_sanitation_20',
+        ]
+        for prop in request_choices:
+            if prop in data:
+                if data[prop] == 1 or data[prop] == '1':
+                    data[prop] = RequestChoices.REQUESTED
+                elif data[prop] == 2 or data[prop] == '2':
+                    data[prop] = RequestChoices.PLANNED
+                elif data[prop] == 3 or data[prop] == '3':
+                    data[prop] = RequestChoices.COMPLETE
+                else:
+                    data[prop] = RequestChoices.NO
+
+        if instance is not None:
+            serializer = CreateFieldReportSerializer(instance, data=data)
+        else:
+            serializer = CreateFieldReportSerializer(data=data)
+        return serializer
+
+    def map_foreign_key_relations(self, data):
+        # The request data object will come with a lot of relation mappings.
+        # For foreign key, we want to replace instance ID's with querysets.
+
+        # Query foreign key relations, these are attached on model save/update.
+        mappings = [
+            ('user', User),
+            ('dtype', DisasterType),
+            ('event', Event),
+        ]
+        for (prop, model) in mappings:
+            if prop in data and data[prop] is not None:
+                try:
+                    data[prop] = model.objects.get(pk=data[prop])
+                except:
+                    raise BadRequest('Valid %s is required' % prop)
+            elif prop is not 'event':
+                raise BadRequest('Valid %s is required' % prop)
+
+        return data
+
+    def map_many_to_many_relations(self, data):
+        # Query many-to-many mappings. These are removed from the data object,
+        # So they can be added later.
+        mappings = [
+            ('countries', Country),
+            ('regions', Region),
+            ('districts', District),
+        ]
+
+        locations = {}
+        for (prop, model) in mappings:
+            if prop in data and hasattr(data[prop], '__iter__') and len(data[prop]):
+                locations[prop] = list(data[prop])
+            if prop in data:
+                del data[prop]
+
+        # Sources, actions, and contacts
+        mappings = [
+            ('actions_taken'),
+            ('contacts'),
+            ('sources'),
+        ]
+
+        meta = {}
+        for (prop) in mappings:
+            if prop in data and hasattr(data[prop], '__iter__') and len(data[prop]):
+                meta[prop] = list(data[prop])
+            if prop in data:
+                del data[prop]
+
+        return data, locations, meta
+
+    def save_locations(self, instance, locations, is_update=False):
+        if is_update:
+            instance.districts.clear()
+            instance.countries.clear()
+            instance.regions.clear()
+        if 'districts' in locations:
+            instance.districts.add(*locations['districts'])
+        if 'countries' in locations:
+            instance.countries.add(*locations['countries'])
+            # Add countries in automatically, based on regions
+            countries = Country.objects.filter(pk__in=locations['countries'])
+            instance.regions.add(*[country.region for country in countries if (
+                country.region is not None
+            )])
+
+    def save_meta(self, fieldreport, meta, is_update=False):
+        if is_update:
+            ActionsTaken.objects.filter(field_report=fieldreport).delete()
+            FieldReportContact.objects.filter(field_report=fieldreport).delete()
+            Source.objects.filter(field_report=fieldreport).delete()
+
+        if 'actions_taken' in meta:
+            for action in meta['actions_taken']:
+                actions = action['actions']
+                del action['actions']
+                actions_taken = ActionsTaken.objects.create(field_report=fieldreport, **action)
+                actions_taken.actions.add(*actions)
+
+        if 'contacts' in meta:
+            FieldReportContact.objects.bulk_create(
+                [FieldReportContact(field_report=fieldreport, **fields) for fields in meta['contacts']]
+            )
+
+        if 'sources' in meta:
+            for source in meta['sources']:
+                stype, created = SourceType.objects.get_or_create(name=source['stype'])
+                source['stype'] = stype
+            Source.objects.bulk_create(
+                [Source(field_report=fieldreport, **fields) for fields in meta['sources']]
+            )
+
+class CreateFieldReport(CreateAPIView, GenericFieldReportView):
+    authentication_classes = (TokenAuthentication,)
+    permission_classes = (IsAuthenticated,)
+    queryset = FieldReport.objects.all()
+    serializer_class = CreateFieldReportSerializer
+
+    def create_event(self, report):
+        event = Event.objects.create(
+            name=report.summary,
+            dtype=report.dtype,
+            disaster_start_date=report.created_at,
+            auto_generated=True,
+            auto_generated_source=SOURCES['new_report'],
+        )
+        report.event = event
+        report.save()
+        return event
+
+    def create(self, request):
+        serializer = self.serialize(request.data)
+        if not serializer.is_valid():
+            raise BadRequest(serializer.errors)
+
+        data = self.map_foreign_key_relations(request.data)
+        data, locations, meta = self.map_many_to_many_relations(data)
+
+        try:
+            fieldreport = FieldReport.objects.create(**data)
+        except:
+            raise BadRequest('Could not create field report')
+
+        ### Creating relations ###
+        # These are *not* handled in a transaction block.
+        # The data model for these is very permissive. We're more interested in the
+        # Numerical data being there than not.
+        errors = []
+        try:
+            self.save_locations(fieldreport, locations)
+        except Exception as e:
+            errors.append(e)
+
+        try:
+            self.save_meta(fieldreport, meta)
+        except Exception as e:
+            errors.append(e)
+
+        # If the report doesn't have an emergency attached, create one.
+        if fieldreport.event is None:
+            event = self.create_event(fieldreport)
+            try:
+                self.save_locations(event, locations)
+            except Exception as e:
+                errors.append(e)
+
+        if len(errors):
+            logger.error('%s errors creating new field reports' % len(errors))
+            for error in errors:
+                logger.error(str(error)[:200])
+
+        return Response({'id': fieldreport.id}, status=HTTP_201_CREATED)
+
+class UpdateFieldReport(UpdateAPIView, GenericFieldReportView):
+    authentication_classes = (TokenAuthentication,)
+    permission_classes = (IsAuthenticated,)
+    queryset = FieldReport.objects.all()
+    serializer_class = CreateFieldReportSerializer
+
+    def partial_update(self, request, *args, **kwargs):
+        self.update(request, *args, **kwargs)
+
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+        serializer = self.serialize(request.data, instance=instance)
+        if not serializer.is_valid():
+            raise BadRequest(serializer.errors)
+
+        data = self.map_foreign_key_relations(request.data)
+        data, locations, meta = self.map_many_to_many_relations(data)
+
+        try:
+            serializer.save()
+        except Exception as e:
+            raise BadRequest('Could not update field report')
+
+        errors = []
+        try:
+            self.save_locations(instance, locations, is_update=True)
+        except Exception as e:
+            errors.append(e)
+
+        try:
+            self.save_meta(instance, meta, is_update=True)
+        except Exception as e:
+            errors.append(e)
+
+        if len(errors):
+            logger.error('%s errors creating new field reports' % len(errors))
+            for error in errors:
+                logger.error(str(error)[:200])
+
+        return Response({'id': instance.id}, status=HTTP_200_OK)