+++ notifications/hello.py
@@ -1,0 +1,14 @@
+import random
+hellos = [
+    'Hey there',
+    'Hello!',
+    'G\'day',
+    'Howdy',
+    'Oh hi',
+    'Hi!',
+    'How are you?',
+    ':waves:',
+]
+
+def get_hello():
+    return random.choice(hellos)